Fix some

+++ flowers/fetch_image_embeddings.py
@@ -5,14 +5,20 @@ import logging
 from subprocess import call
 from multiprocessing import Pool
 
-from trainer.emb import get_shard, id_to_path
+from tensorflow.python.lib.io import file_io
 
-IMAGES_COUNT_COL = 3
+from trainer.emb import get_shard, id_to_path, ID_COL, IMAGES_COUNT_COL
+
 emb_path = 'data/image_embeddings'
+gs_emb_path = 'gs://towneers-ml/article_classifier'
 
 def remove(id):
     filepath = "%s/%s" % (emb_path, id_to_path(id))
-    return os.remove(filepath)
+    gs_filepath = '%s/%s' % (gs_emb_path, filepath)
+    if file_io.file_exists(gs_filepath):
+        file_io.delete_file(gs_filepath)
+    if os.path.exists(filepath):
+        return os.remove(filepath)
 
 def down(id):
     shard = get_shard(id)
@@ -23,24 +29,39 @@ def down(id):
         except OSError as e:
             if e.errno != 17: # File exists
                 raise e
-    to_path = "%s/%d.emb" % (to_path, id)
+    to_filepath = "%s/%d.emb" % (to_path, id)
     url = 'http://ml.daangn.com/articles/image_embeddings/%s' % id_to_path(id)
     logging.info('down: %s', url)
-    return call(['curl', '-f', '--connect-timeout', '2', '-o', to_path, url])
+    result = call(['curl', '-f', '--connect-timeout', '2', '-o', to_filepath, url])
+    if not os.path.exists(to_filepath):
+        return result
+
+    to_gs_filepath = '%s/%s' % (gs_emb_path, to_filepath)
+    if file_io.file_exists(to_gs_filepath):
+        return result
+
+    to_gs_path = '%s/%s' % (gs_emb_path, to_path)
+    if not file_io.is_directory(to_gs_path):
+        file_io.create_dir(to_gs_path)
+    file_io.copy(to_filepath, to_gs_filepath)
+    return result
 
 def main():
     with open('data/emb.csv') as f:
         rows = [line.split(',') for line in f.readlines()[1:]]
-        new_ids = set([int(row[0]) for row in rows if row[IMAGES_COUNT_COL] != '0'])
+        new_ids = set([int(row[ID_COL]) for row in rows if row[IMAGES_COUNT_COL] != '0'])
 
     pathes = glob.glob('%s/*/*.emb' % emb_path)
     local_ids = set([int(x.split('/')[-1][:-4]) for x in pathes])
 
-    print "local ids count: %d" % len(local_ids)
-    print "new ids count: %d" % len(new_ids)
+    if len(new_ids) < 1:
+        raise Exception('new_ids is zero')
 
     add_ids = new_ids - local_ids
     remove_ids = local_ids - new_ids
+
+    print "add ids count: %d" % len(add_ids)
+    print "remove ids count: %d" % len(remove_ids)
 
     p = Pool()
